+++ CHANGELOG.md
@@ -1,3 +1,8 @@
+# 0.1.1 (xxxx-xx-xx)
+
+- [Fixed PDF export on Windows](https://github.com/DigiScore/neoscore/issues/37)
+- Updated `neoscore.render_image()` to block by default. Set the new kwarg `wait=False` to finalize export asynchronously.
+
 # 0.1.0 (2020-05-15)
 
 Initial release! ❤️❤️❤️